update tests to cover more methods of TaskManager interface
