+++ src/ru/yandex/service/TaskManager.java
@@ -4,6 +4,7 @@ import ru.yandex.model.Epic;
 import ru.yandex.model.Subtask;
 import ru.yandex.model.Task;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 public interface TaskManager {
@@ -48,4 +49,11 @@ public interface TaskManager {
     List<Task> getPrioritizedTasks();
 
     List<Task> getHistory();
+
+    static boolean checkOverlap(Task task1, Task task2) {
+        LocalDateTime start1 = task1.getStartTime(), start2 = task2.getStartTime(), end1 = task1.getEndTime(), end2
+                = task2.getEndTime();
+        return ((!start1.isBefore(start2) && start1.isBefore(end2)) || (!start2.isBefore(start1) && start2.isBefore(
+                end1)));
+    }
 }
